requirements.txt
# AURA — Python Dependencies
# Install: pip install -r requirements.txt
# LeRobot (vendored): pip install -e lerobot
# Python: 3.11+ (conda recommended)

# ── Web framework ──────────────────────────────────────────────
fastapi>=0.115.0,<1.0.0
uvicorn>=0.30.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0

# ── Data & scientific computing ────────────────────────────────
numpy>=1.26.0,<2.0.0
pandas>=2.2.0,<2.4.0

# ── Deep learning ──────────────────────────────────────────────
torch>=2.2.1,<2.8.0
torchvision>=0.21.0,<0.23.0

# ── LLM / vision-language ─────────────────────────────────────
transformers>=4.57.0,<5.0.0
google-generativeai>=0.8.0,<1.0.0

# ── RL ─────────────────────────────────────────────────────────
gymnasium>=1.1.1,<2.0.0

# ── Computer vision ────────────────────────────────────────────
opencv-python>=4.9.0,<4.13.0

# ── Hardware communication ─────────────────────────────────────
pyserial>=3.5,<4.0
python-can>=4.4.0,<5.0.0

# ── Configuration ──────────────────────────────────────────────
pyyaml>=6.0,<7.0

# ── Development ────────────────────────────────────────────────
pytest>=8.1.0,<9.0.0
ruff>=0.6.0,<1.0.0

# ── LeRobot extras (install after: pip install -e lerobot) ────
# Motor SDKs (needed for hardware):
#   pip install "lerobot[dynamixel,feetech,intelrealsense]"
# Policy extras (needed for SmolVLA, HIL-SERL):
#   pip install "lerobot[smolvla,hilserl]"
